[MODEXPW-528] Update unit tests
