+++ src/test/java/org/folio/dew/batch/acquisitions/edifact/jobs/ExportHistoryTaskletTest.java
@@ -12,6 +12,7 @@ import org.springframework.batch.core.JobParametersBuilder;
 import org.springframework.batch.item.ExecutionContext;
 import org.springframework.batch.test.JobLauncherTestUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.annotation.DirtiesContext;
 
@@ -26,6 +27,7 @@ import static org.mockito.Mockito.doReturn;
 
 class ExportHistoryTaskletTest extends BaseBatchTest {
   @Autowired
+  @Qualifier("edifactOrdersExportJob")
   private Job edifactExportJob;
   @MockBean
   private OrganizationsService organizationsService;
